# web/backend/api/docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: "https://api.eesast.com"
    description: Remote API Server
  - url: "http://localhost:{port}"
    description: Local API Server
    variables:
      port:
        description: 端口号
        default: "28888"
info:
  title: API Backend for EESAST
  description: |
    清华大学电子工程系学生科协后端 API。了解更多 EESAST 内容，请访问：[GitHub 主页](https://github.com/eesast/)或 [EESAST 官方网站](https://eesast.com/)。
  version: 1.0.0
  contact:
    name: API Support
    email: i@raineggplant.com
...
paths:
...
  "/v1/announcements":
    get:
      tags:
        - announcements
      summary: 返回匹配查询的公告
      description: 可以通过 query 进行筛选
      operationId: findAnnouncements
      parameters:
        - name: contestId
          in: query
          schema:
            type: number
          description: 比赛ID
        - name: priority
          in: query
          schema:
            type: number
          description: 公告显示优先级
        - name: begin
          in: query
          schema:
            type: number
          description: 按时间降序排列，起始公告的位置
        - name: end
          in: query
          schema:
            type: number
          description: 按时间降序排列，结束公告的位置
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          description: Server error

    post:
      tags:
        - announcements
      summary: 发布一篇新的公告
      description: 需要 `root` 或 `organizer` 角色
      operationId: addAnnouncements
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/addAnnouncement"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新发布公告的 URL
              schema:
                type: string
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Server error
  "/v1/announcements/{announcementId}":
    get:
      tags:
        - announcements
      summary: 根据 ID 获取公告
      operationId: getAnnouncementById
      parameters:
        - name: announcementId
          description: 公告的 ID
          in: path
          required: true
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: Announcement does not exist"
        500:
          description: Server error

    put:
      tags:
        - announcements
      summary: 更新已经存在的公告
      description: 需要 `root` 或 `organizer` 角色
      operationId: updateAnnouncement
      security:
        - bearerAuth: []
      parameters:
        - name: announcementId
          in: path
          description: 要更新的公告ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/updateAnnouncement"
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新公告的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Announcement does not exist"
        500:
          description: Server error

    delete:
      tags:
        - announcements
      summary: 删除一篇公告
      description: 需要 `root` 或 `organizer` 角色
      operationId: deleteAnnouncement
      security:
        - bearerAuth: []
      parameters:
        - name: announcementId
          in: path
          description: 要删除的公告ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Announcement does not exist"
...
components:
  schemas:
...
    Date:
      type: string
      format: date-time
      description: >-
        日期对象。使用时采用 ISO
        格式字符串表示。[更多信息](https://www.iso.org/iso-8601-date-and-time-format.html)
...
    Announcement:
      type: object
      required:
        - id
        - contestId
        - title
        - content
        - priority
      properties:
        id:
          type: number
          description: 自增值
        contestId:
          type: number
          description: 比赛ID
        title:
          type: string
          description: 公告标题
        content:
          type: string
          description: 公告内容
        priority:
          type: number
          default: 1
          description: 公告显示优先级
        createdAt:
          $ref: "#/components/schemas/Date"
        createdBy:
          type: number
        updatedAt:
          $ref: "#/components/schemas/Date"
        updatedBy:
          type: number
...
  requestBodies:
...
    addAnnouncement:
      content:
        application/json:
          schema:
            type: object
            required:
              - contestId
              - title
              - content
            properties:
              contestId:
                type: number
                description: 比赛ID
              title:
                type: string
                description: 公告标题
              content:
                type: string
                description: 文章内容(支持 Markdown)
              priority:
                type: number
                description: 优先级，用于排序
      description: 发布公告的提交格式
      required: true
    updateAnnouncement:
      content:
        application/json:
          schema:
            type: object
            properties:
              contestId:
                type: number
                description: 比赛ID
              title:
                type: string
                description: 公告标题
              content:
                type: string
                description: 公告内容(支持 Markdown)
              priority:
                type: number
                description: 优先级，用于排序
      description: 更新公告的提交格式
      required: true
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                description: 错误列表
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      description: 错误位置
                      example: ".params.articleId"
                    message:
                      type: string
                      description: 错误信息
                      example: "should be number"
                    errorCode:
                      type: string
                      description: 错误码
                      example: "type.openapi.validation"
      description: 包含请求错误信息的 JSON
    UnauthorizedError:
      description: "Access token is missing, invalid or expired / Permission denied"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT